+++ src/Normalization/FinalEolNormalization.php
@@ -5,18 +5,19 @@ declare(strict_types=1);
 namespace Normalizator\Normalization;
 
 use Normalizator\Attribute\Normalization;
-use Normalizator\Configuration\Configuration;
 use Normalizator\EventDispatcher\Event\NormalizationEvent;
 use Normalizator\EventDispatcher\EventDispatcher;
 use Normalizator\Filter\FilterManager;
 use Normalizator\Finder\File;
 use Normalizator\Util\EolDiscovery;
 
+use function Normalizator\preg_match_all;
+
 /**
  * Normalizes final newlines.
  */
 #[Normalization(
-    name: 'final-eol',
+    name: 'final_eol',
     filters: [
         'file',
         'plain-text',
@@ -33,27 +34,32 @@ class FinalEolNormalization implements NormalizationInterface, ConfigurableNorma
      */
     public const MAX_EXTRA_FINAL_EOLS = 1;
 
-    /**
-     * Overridden configuration values.
-     *
-     * @var array<string,mixed>
-     */
-    private array $overrides;
+    private string $eol = EolDiscovery::DEFAULT_EOL;
+    private int $maxExtraFinalEols = self::MAX_EXTRA_FINAL_EOLS;
 
     /**
      * Class constructor.
      */
     public function __construct(
-        private Configuration $configuration,
         private FilterManager $filterManager,
         private EventDispatcher $eventDispatcher,
         private EolDiscovery $eolDiscovery
     ) {
     }
 
-    public function configure(array $values): void
+    public function configure(mixed ...$options): void
     {
-        $this->overrides = $values;
+        if (isset($options['eol']) && is_string($options['eol'])) {
+            $this->eol = $options['eol'];
+        } else {
+            $this->eol = EolDiscovery::DEFAULT_EOL;
+        }
+
+        if (isset($options['max_extra_final_eols']) && is_int($options['max_extra_final_eols'])) {
+            $this->maxExtraFinalEols = $options['max_extra_final_eols'];
+        } else {
+            $this->maxExtraFinalEols = self::MAX_EXTRA_FINAL_EOLS;
+        }
     }
 
     /**
@@ -71,7 +77,7 @@ class FinalEolNormalization implements NormalizationInterface, ConfigurableNorma
         $newlines = $this->getFinalEols($content);
         $trimmed = rtrim($content, "\r\n");
 
-        $max = $this->getConfig('max_extra_final_eols', self::MAX_EXTRA_FINAL_EOLS);
+        $max = $this->maxExtraFinalEols;
 
         // Empty content doesn't need one final newline when max = 1
         if ('' === $trimmed && 1 === $max) {
@@ -87,8 +93,12 @@ class FinalEolNormalization implements NormalizationInterface, ConfigurableNorma
         }
 
         // Then insert one missing final EOL if not present yet.
-        if ('' !== $trimmed && !in_array(substr($trimmed, -1), ["\n", "\r"], true)) {
-            $trimmed .= $this->eolDiscovery->getPrevailingEol($content);
+        if (
+            '' !== $trimmed
+            && !in_array(substr($trimmed, -1), ["\n", "\r"], true)
+            && 0 < $max
+        ) {
+            $trimmed .= $this->getPrevailingEol($content, $file);
         }
 
         if ($content !== $trimmed) {
@@ -97,15 +107,6 @@ class FinalEolNormalization implements NormalizationInterface, ConfigurableNorma
         }
 
         return $file;
-    }
-
-    private function getConfig(string $key, mixed $default = null): mixed
-    {
-        if (isset($this->overrides[$key])) {
-            return $this->overrides[$key];
-        }
-
-        return $this->configuration->get($key, $default);
     }
 
     /**
@@ -143,5 +144,24 @@ class FinalEolNormalization implements NormalizationInterface, ConfigurableNorma
         }
 
         return array_reverse($newlines);
+    }
+
+    /**
+     * Get EOL based on the prevailing LF, CRLF or CR newline characters.
+     */
+    private function getPrevailingEol(string $content, File $file): string
+    {
+        // Match all LF, CRLF and CR EOL characters
+        preg_match_all('/(*BSR_ANYCRLF)\R/', $content, $matches);
+
+        if (is_array($matches[0])) {
+            $counts = array_count_values($matches[0]);
+            arsort($counts);
+
+            return (string) key($counts);
+        }
+
+        // For single line files the EOL needs to be determined.
+        return $this->eolDiscovery->getEolForFile($file, $this->eol);
     }
 }