+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Command line options validation and resolving configuration.
 - Finder iterator counting.
 - Docker tag format fixed.